python/restful-booker-tests/pyproject.toml
[project]
name = "restful-booker-tests"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
